Fix: Change parameter types of div function to double for accurate division

+++ 01-basics/Practicing Functions/Multifiles 1/arithmetic.cpp
@@ -3,7 +3,7 @@ int add(int x, int y, int z);
 
 int mul(int a, int b, int c);
 
-double div(int p, int q, int r);
+double div(double p, double q, double r);
 
 int sub(int u, int v, int w);
